Revert "Revert "Updates the plugin for Elgg 1.9+""
