+++ README.md
@@ -1,4 +1,14 @@
-= galliElggEmbed =
+# galliElggEmbed
+
+## Features
+
+ * Allows users to embed their Elgg pages in other websites
+
+## Requirements
+
+ * Elgg 1.9 or above
+
+## Info
 
  *	Author : Mahin Akbar | Team Webgalli
  *	Team Webgalli | Elgg developers and consultants
@@ -7,11 +17,3 @@
  *	@package galliElggEmbed
  *	Licence : GPLV2
  *	Copyright : Team Webgalli 2013-2015
- 
-= Feature =
-
- * Allows users to embed their Elgg pages in other websites.
- 
-= Requirements =
-
- * Elgg 1.8 and above.
