+++ composer.json
@@ -1,0 +1,8 @@
+{
+    "name": "teamwebgalli/galliElggEmbed",
+    "type": "elgg-plugin",
+    "require": {
+        "composer/installers": ">=1.0.8"
+    },
+    "license": "GPL-2.0"
+}
