+++ manifest.xml
@@ -2,7 +2,7 @@
 <plugin_manifest xmlns="http://www.elgg.org/plugin_manifest/1.8">
 	<name>galliElggEmbed</name>
 	<author>Mahin Akbar | Team Webgalli</author>
-	<version>2.0</version>
+	<version>3.0</version>
 	<category>Team Webgalli</category>
 	<category>social</category>
 	<description>Embed elgg pages on other websites.</description>
@@ -12,6 +12,6 @@
 	<license>GNU General Public License version 2</license>
 	<requires>
 		<type>elgg_release</type>
-		<version>1.0</version>
+		<version>1.9</version>
 	</requires>
 </plugin_manifest>
